performing a security audit
This is an activity where an audit is performed for a fictional company called "Botium Toys". In reality, it’s an activity provided by the Google Cybersecurity certificate.

Description:

This scenario is based on a fictional company:

Botium Toys is a small U.S.-based company that develops and sells toys. The company has a single physical location, which serves as its main office, storefront, and warehouse for its products. However, Botium Toys' online presence has grown, attracting customers both in the U.S. and internationally. As a result, its IT department is under increasing pressure to support its global online market.

The IT department head has decided that an internal IT audit is necessary. She is concerned about maintaining compliance and business operations as the company grows without a clear plan. She believes that an internal audit can help better protect the company's infrastructure and assist in identifying and mitigating potential risks, threats, or vulnerabilities to critical assets. The executive is also interested in ensuring compliance with regulations related to internal processing, online payment acceptance, and doing business in the European Union (EU).

The IT manager starts by applying the National Institute of Standards and Technology (NIST CSF) Cybersecurity Framework, defining the scope and goals of the audit, listing the assets currently managed by the IT department, and completing a risk assessment. The goal of the audit is to provide an overview of the risks and/or fines the company might face due to the current state of its security posture.

Your task is to review the scope, objectives, and risk assessment report provided by the IT manager. Then, perform an internal audit by completing a checklist of controls and compliance.