Conducting a security audit
Internal IT Audit for Botium Toys (Fictitious Company)

This activity involves an internal IT audit for Botium Toys, a fictitious company that sells toys both nationally and internationally. The audit follows the NIST Cybersecurity Framework to identify potential risks, vulnerabilities, and threats in the company's infrastructure.

It includes an assessment of managed assets, evaluation of online operations and payment systems, and compliance with regulations such as the EU’s GDPR. Recommendations were made to improve security posture and ensure compliance in a global business environment.